Showing recent match card component in shared module.
